refactor: 가게 정보 카드 데이터 패칭 위치 변경 [TOKTOT-412]
