--- src/entities/review/ui/ReviewStore.tsx
@@ -1,5 +1,0 @@
-const ReviewStore = () => {
-	return <div>ReviewStore</div>;
-};
-
-export default ReviewStore;
